Add basic iteration entry module (#30) (#31)
* Create module template for iterations input

* Relocate data entry UI components

The fluidPage() code for entering parameters now lives directly in the app_ui() rather than behind the data entry module. This makes it easier to insert other modules into the page such as iterations.

* Create functioning module for entering iterations

Iterations can be entered in a side panel on the data entry page, and are currently printed in the testing portion of the stats calculations server (just to demonstrate successful passing of values).