Various updates to calculations and GH Actions for deployment (#64)
* Change triggers for running R CMD checks

Run checks only on pushes (i.e. when a branch is merged--this necessarily creates a push) to the master and development branches. Enable workflow dispatching for any branch.

* Don't autorun on push to development

* Set up ShinyApps io deployment action (#60)

Deploys only the master branch. Consider a development deployment in the future

* Install libgit2-dev in action workflow

* Fix typo in action

* Add account name attempt

* Update renv dependencies for action

* Remove browser call

* Force updates to Shiny site

* Use golem fix for deploying to shinyapps.io

* Add spinners to loadable components

* Ensure evaluation of T1 error, T2 error/power calcs

This forces all of them to be done at once rather than waiting for tabs (where they are hidden) to be pressed. Has downstream effect of making the calculations immediately available for the calculations on the distributions page.

* Remove commented-out code

* Include CI for Type II error