Development (#67)
* Vignettes (#55)

* Create empty vignette

* Include basic notes on desired vignette structure

* Add space for setting app options

* Write more descriptive vignette and add app options

There is now more content describing the purpose and use of the application. Options for pre-setting information on the app are now being developed

* Finish vignette draft

Include example data for working with output. Ignore intermediate/testing output for checking whether vignettes can be render (doc and Meta).

* Update renv lockfile

* Remove pkgload from project Imports

This package is only needed for using the load_all() shim in app.R, but this script exists for deployment on shinyapps.io and the image is built using GH Actions which activates {renv}. {pkgload} is captured in the renv lockfile so declaring it as an Import is redundant.

* Expand vignette examples

Make misc updates to documentation too.

* Update README with more useful information

Add progress bars to template code in vignette.

* Create {pkgdown} formatted website

* Rebuild website only on main push or workflow dispatch